driver: Fix warnings (#5)
There are several trivial warnings that appear when building. Address
them as necessary by fixing the problems. In general, unused parameters
to functions that are part of an interface cannot be removed so they are
marked with the UNUSED macro. Signed comparisons can never be negative
and uints can never be larger than UINT_MAX. XAIE columns are generally
a u8 type (limited to 256 columns) although stored in a u32 in places.
If we ever need to support more than 256 columns this needs to change
across the board.

Signed-off-by: Joel Nider <joel.nider@amd.com>